-add searching methods in equipment, cross origin in weather -

+++ backend/src/main/java/com/example/skicurort/equipment/EquipmentService.java
@@ -1,15 +1,14 @@
 package com.example.skicurort.equipment;
 
-import java.util.List;
-import java.util.NoSuchElementException;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 import static com.example.skicurort.equipment.EquipmentMapper.mapToDTO;
 import static com.example.skicurort.equipment.EquipmentMapper.mapToDTOs;
 import static com.example.skicurort.equipment.EquipmentMapper.mapToEntity;
+
+import java.util.List;
+import java.util.NoSuchElementException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
@@ -29,6 +28,12 @@ public class EquipmentService {
   List<EquipmentDTO> getAllEquipmentByType(String type) {
     return mapToDTOs(equipmentRepository.findAllByType(type));
   }
+  List<EquipmentDTO> getEquipmentByTypeAndSize(String type,String size){
+    return mapToDTOs(equipmentRepository.findAllByTypeAndSize(type,size));
+  }
+  List<EquipmentDTO> getEquipmentByTypeAndSizeAndAvailable(String type,String size,boolean available){
+    return mapToDTOs(equipmentRepository.findAllByTypeAndSizeAndAvailable(type,size,available));
+  }
 
   EquipmentDTO getOneItem(Long id) {
     return mapToDTO(
@@ -37,12 +42,11 @@ public class EquipmentService {
             .orElseThrow(() -> new NoSuchElementException("item with id: " + id + " not exist")));
   }
 
-  EquipmentDTO addEquipment( EquipmentDTO equipmentDTO) {
+  EquipmentDTO addEquipment(EquipmentDTO equipmentDTO) {
     return mapToDTO(equipmentRepository.save(mapToEntity(equipmentDTO)));
   }
 
-  EquipmentDTO editEquipment( Long id, EquipmentDTO equipmentDTO)
-      throws NoSuchElementException {
+  EquipmentDTO editEquipment(EquipmentDTO equipmentDTO, Long id) throws NoSuchElementException {
     Equipment equipment =
         equipmentRepository
             .findById(id)
@@ -54,11 +58,10 @@ public class EquipmentService {
     return mapToDTO(equipmentRepository.save(equipment));
   }
 
-  void deleteEquipment(Long id) {
-    Equipment equipment =
-        equipmentRepository
-            .findById(id)
-            .orElseThrow(() -> new NoSuchElementException("item with id: " + id + " not exist"));
+  void deleteEquipment(Long id) throws NoSuchElementException {
+    if (!equipmentRepository.existsById(id)) {
+      throw new NoSuchElementException();
+    }
     equipmentRepository.deleteById(id);
   }
 }
